Optimized facade mixing a little more.
Ideally it seems like we could drop hasOwnProperty here, since
we have a whitelist, but if(o in p) without the hasOwnProperty is
just as slow in Chrome 21 as the current implementation, if not slower
(which kind of makes sense to me - I would expect if(o.hasOwnProperty(p))
to be faster than if(p in o) in theory - would be surprised if that optimization
shows up in other engines.

So, anyway, going with this approach which is the only one which improves on the
current implementation across all targeted browsers.

http://jsperf.com/hop-in/12

NOTE: I saw the same results in native Chrome profiling also.
(cherry picked from commit 627216ec336d9e01a26cfb30f6d0a22bbea8ef3e)