AUI-2062 Don't prevent-default in 'on' for document node
The _setMsTouchActions method modifies node behavior during touch events and prevents the default response from the browser.

This will prevent pages from being scrolled in any direction if there is _any_ event listener set on the document node. For example in https://github.com/yui/yui3/issues/1343 DDM sets an 'ongesturemove' listener on the document  in order to know when to update the locations of dragged elements. BUT, the document can no longer be scrolled via touch because of the side-effect in _setTouchActions.

This fixes #1343 and maybe #1795